% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glueko-package.R
\docType{package}
\name{glueko-package}
\alias{glueko}
\alias{glueko-package}
\title{glueko: Glue Package Between Tidytext And Rcppmecab for Korean}
\description{
Helper functions to use tidytext with RcppMeCab for Korean nlp.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mrchypark/glueko}
  \item Report bugs at \url{https://github.com/mrchypark/glueko/issues}
}

}
\author{
\strong{Maintainer}: Chanyub Park \email{mrchypark@gmail.com} (0000-0001-6474-2570)

}
\keyword{internal}
